+++ app/actions/bundle.actions.js
@@ -1,13 +1,18 @@
 import sort from 'fast-sort';
+import traverse from 'traverse';
 import { bundleConstants } from '../constants/bundle.constants';
 import { bundleService } from '../services/bundle.service';
+import { dblDotLocalConfig } from '../constants/dblDotLocal.constants';
 
 export const bundleActions = {
   mockFetchAll,
   fetchAll,
   delete: removeBundle,
+  setupBundlesEventSource,
+  downloadResources,
+  requestSaveBundleTo,
   toggleModePauseResume,
-  toggleSelectBundle,
+  toggleSelectBundle
 };
 
 export default bundleActions;
@@ -15,13 +20,10 @@ export default bundleActions;
 export function mockFetchAll() {
   return dispatch => {
     dispatch(request());
-    return new Promise((resolve) => {
+    return new Promise(resolve => {
       const mockBundles = getMockBundles();
       resolve(mockBundles);
-    }).then(
-      bundles => dispatch(success(bundles)),
-      error => dispatch(failure(error))
-    );
+    }).then(bundles => dispatch(success(bundles)), error => dispatch(failure(error)));
   };
 
   function request() {
@@ -41,10 +43,7 @@ export function fetchAll() {
 
     return bundleService
       .fetchAll()
-      .then(
-        bundles => dispatch(success(bundles)),
-        error => dispatch(failure(error))
-      );
+      .then(bundles => dispatch(success(bundles)), error => dispatch(failure(error)));
   };
 
   function request() {
@@ -55,6 +54,85 @@ export function fetchAll() {
   }
   function failure(error) {
     return { type: bundleConstants.FETCH_FAILURE, error };
+  }
+}
+
+export function setupBundlesEventSource(authentication) {
+  return dispatch => {
+    console.log('SSE connect to Bundles');
+    const eventSource = new EventSource(`${dblDotLocalConfig.getHttpDblDotLocalBaseUrl()}/events/${authentication.user.auth_token}`);
+    eventSource.onmessage = (event) => {
+      console.log(event);
+    };
+    eventSource.onopen = () => {
+      console.log('Connection to event source opened.');
+    };
+    eventSource.onerror = (error) => {
+      console.log('EventSource failed.');
+      console.log(error);
+    };
+    const listeners = {
+      'storer/execute_task': listenStorerExecuteTaskDownloadResources,
+      'downloader/receiver': listenDownloaderReceiver,
+      'downloader/status': (e) => listenDownloaderStatus(e, dispatch),
+      'storer/update_from_download': listenStorerUpdateFromDownload,
+    };
+    Object.keys(listeners).forEach((evType) => {
+      const handler = listeners[evType];
+      eventSource.addEventListener(evType, handler);
+    });
+  };
+
+  function listenStorerExecuteTaskDownloadResources(e) {
+    console.log(e);
+  }
+
+  function listenDownloaderReceiver(e) {
+    console.log(e);
+  }
+
+  /* downloader/status
+   * {'event': 'downloader/status', 'data': {'args': ('48a8e8fe-76ac-45d6-9b3a-d7d99ead7224', 4, 8), 'component': 'downloader', 'type': 'status'}}
+   */
+  function listenDownloaderStatus(e, dispatch) {
+    console.log(e);
+    const data = JSON.parse(e.data);
+    const bundleId = data.args[0];
+    const resourcesDownloaded = data.args[1];
+    const resourcesToDownload = data.args[2];
+    dispatch(updateDownloadStatus(bundleId, resourcesDownloaded, resourcesToDownload));
+  }
+
+  function updateDownloadStatus(_id, resourcesDownloaded, resourcesToDownload) {
+    return {
+      type: bundleConstants.DOWNLOAD_RESOURCES_UPDATED,
+      id: _id,
+      resourcesDownloaded,
+      resourcesToDownload
+    };
+  }
+
+  function listenStorerUpdateFromDownload(e) {
+    console.log(e);
+  }
+}
+
+export function downloadResources(id) {
+  return async dispatch => {
+    try {
+      const manifestResourcePaths = await bundleService.getManifestResourcePaths(id);
+      manifestResourcePaths.unshift('metadata.xml');
+      dispatch(request(id, manifestResourcePaths));
+      await bundleService.downloadResources(id);
+    } catch (error) {
+      dispatch(failure(id, error));
+    }
+  };
+  function request(_id, manifestResourcePaths) {
+    return { type: bundleConstants.DOWNLOAD_RESOURCES_REQUEST, id: _id, manifestResourcePaths };
+  }
+  function failure(_id, error) {
+    return { type: bundleConstants.DOWNLOAD_RESOURCES_FAILURE, id, error };
   }
 }
 
@@ -85,6 +163,85 @@ function removeBundle(id) {
   }
 }
 
+export function requestSaveBundleTo(id, selectedFolder) {
+  return async dispatch => {
+    const bundleInfo = await bundleService.fetchById(id);
+    const bundleBytesToSave = traverse(bundleInfo.store.file_info).reduce(addByteSize, 0);
+    const resourcePaths = await bundleService.getResourcePaths(id);
+    resourcePaths.unshift('metadata.xml');
+    const resourcePathsProgress = resourcePaths.reduce((acc, resourcePath) => {
+      acc[resourcePath] = 0;
+      return acc;
+    }, {});
+    let bundleBytesSaved = 0;
+    dispatch(request(id, selectedFolder, bundleBytesToSave, resourcePaths));
+    resourcePaths.forEach(async resourcePath => {
+      try {
+        const downloadItem = await bundleService.requestSaveResourceTo(
+          selectedFolder,
+          id,
+          resourcePath,
+          (resourceTotalBytesSaved, resourceProgress) => {
+            const originalResourceBytesTransferred = resourcePathsProgress[resourcePath];
+            resourcePathsProgress[resourcePath] = resourceTotalBytesSaved;
+            const bytesDiff = resourceTotalBytesSaved - originalResourceBytesTransferred;
+            bundleBytesSaved += bytesDiff;
+            if (resourceProgress && resourceProgress % 100 === 0) {
+              dispatch(updated(
+                id,
+                resourcePath,
+                resourceTotalBytesSaved,
+                bundleBytesSaved,
+                bundleBytesToSave
+              ));
+            }
+          }
+        );
+        return downloadItem;
+      } catch (error) {
+        dispatch(failure(id, error));
+      }
+    });
+  };
+
+  function addByteSize(accBytes, fileInfoNode) {
+    if (fileInfoNode.is_dir || this.isRoot || fileInfoNode.size === undefined) {
+      return accBytes;
+    }
+    return accBytes + fileInfoNode.size;
+  }
+
+  function request(_id, _folderName, bundleBytesToSave, resourcePaths) {
+    return {
+      type: bundleConstants.SAVETO_REQUEST,
+      id: _id,
+      folderName: _folderName,
+      bundleBytesToSave,
+      resourcePaths
+    };
+  }
+
+  function updated(
+    _id,
+    resourcePath,
+    resourceTotalBytesSaved,
+    bundleBytesSaved,
+    bundleBytesToSave
+  ) {
+    return {
+      type: bundleConstants.SAVETO_UPDATED,
+      id,
+      resourcePath,
+      resourceTotalBytesSaved,
+      bundleBytesSaved,
+      bundleBytesToSave
+    };
+  }
+  function failure(_id, error) {
+    return { type: bundleConstants.SAVETO_FAILURE, id: _id, error };
+  }
+}
+
 export function toggleModePauseResume(id) {
   return { type: bundleConstants.TOGGLE_MODE_PAUSE_RESUME, id };
 }
@@ -96,36 +253,92 @@ export function toggleSelectBundle(id) {
 function getMockBundles() {
   const bundles = [
     {
-      id: 'bundle01', name: 'Test Bundle #1', revision: 3, task: 'UPLOAD', status: 'COMPLETED'
+      id: 'bundle01',
+      name: 'Test Bundle #1',
+      revision: 3,
+      task: 'UPLOAD',
+      status: 'COMPLETED'
     },
     {
-      id: 'bundle02', name: 'Another Bundle', revision: 3, task: 'UPLOAD', status: 'UPLOADING', progress: 63, mode: 'PAUSED'
+      id: 'bundle02',
+      name: 'Another Bundle',
+      revision: 3,
+      task: 'UPLOAD',
+      status: 'IN_PROGRESS',
+      progress: 63,
+      mode: 'PAUSED'
     },
     {
-      id: 'bundle03', name: 'Audio Bundle', revision: 52, task: 'DOWNLOAD', status: 'DOWNLOADING', progress: 12, mode: 'RUNNING'
+      id: 'bundle03',
+      name: 'Audio Bundle',
+      revision: 52,
+      task: 'DOWNLOAD',
+      status: 'IN_PROGRESS',
+      progress: 12,
+      mode: 'RUNNING'
     },
     {
-      id: 'bundle04', name: 'Unfinished Bundle', task: 'UPLOAD', status: 'DRAFT'
+      id: 'bundle04',
+      name: 'Unfinished Bundle',
+      task: 'UPLOAD',
+      status: 'DRAFT'
     },
     {
-      id: 'bundle05', name: 'Unfinished Video Bundle', task: 'UPLOAD', status: 'DRAFT'
+      id: 'bundle05',
+      name: 'Unfinished Video Bundle',
+      task: 'UPLOAD',
+      status: 'DRAFT'
     },
     {
-      id: 'bundle06', name: 'DBL Bundle', task: 'DOWNLOAD', status: 'NOT_STARTED'
+      id: 'bundle06',
+      name: 'DBL Bundle',
+      task: 'DOWNLOAD',
+      status: 'NOT_STARTED'
     },
     {
-      id: 'bundle07', name: 'DBL Bundle 3', task: 'DOWNLOAD', status: 'NOT_STARTED'
+      id: 'bundle07',
+      name: 'DBL Bundle 3',
+      task: 'DOWNLOAD',
+      status: 'NOT_STARTED'
     },
     {
-      id: 'bundle08', name: 'Audio Bundle #2', revision: 40, task: 'DOWNLOAD', status: 'COMPLETED'
+      id: 'bundle08',
+      name: 'Audio Bundle #2',
+      revision: 40,
+      task: 'DOWNLOAD',
+      status: 'COMPLETED'
     },
+    {
+      id: 'bundle09',
+      name: 'Audio Bundle #3',
+      revision: 5,
+      task: 'SAVETO',
+      status: 'IN_PROGRESS',
+      progress: 0,
+    },
+    {
+      id: 'bundle10',
+      name: 'Audio Bundle #4',
+      revision: 4,
+      task: 'SAVETO',
+      status: 'IN_PROGRESS',
+      progress: 66,
+    },
+    {
+      id: 'bundle11',
+      name: 'Audio Bundle #5',
+      revision: 5,
+      task: 'SAVETO',
+      status: 'COMPLETED',
+      progress: 100,
+    }
   ];
-  const taskOrder = ['UPLOAD', 'DOWNLOAD'];
-  const statusOrder = ['UPLOADING', 'DOWNLOADING', 'DRAFT', 'COMPLETED', 'NOT_STARTED'];
+  const taskOrder = ['UPLOAD', 'DOWNLOAD', 'SAVETO'];
+  const statusOrder = ['IN_PROGRESS', 'DRAFT', 'COMPLETED', 'NOT_STARTED'];
   const sortedBundles = sort(bundles).asc([
-    (b) => statusOrder.indexOf(b.status),
-    (b) => taskOrder.indexOf(b.task),
-    (b) => b.name,
+    b => statusOrder.indexOf(b.status),
+    b => taskOrder.indexOf(b.task),
+    b => b.name
   ]);
   return sortedBundles;
 }