+++ app/menu.js
@@ -31,6 +31,7 @@ export default class MenuBuilder {
   }
 
   setupDevelopmentEnvironment() {
+    console.log('setupDevelopmentEnvironment');
     this.mainWindow.openDevTools();
     this.mainWindow.webContents.on('context-menu', (e, props) => {
       const { x, y } = props;