+++ app/services/bundle.service.js
@@ -1,11 +1,18 @@
+import path from 'path';
 import { authHeader } from '../helpers';
 import { dblDotLocalConfig } from '../constants/dblDotLocal.constants';
+import download from './download-with-fetch.flow';
+
 
 export const bundleService = {
   create,
   fetchAll,
   fetchById,
   update,
+  getManifestResourcePaths,
+  downloadResources,
+  getResourcePaths,
+  requestSaveResourceTo,
   delete: removeBundle
 };
 export default bundleService;
@@ -13,7 +20,8 @@ export default bundleService;
 const BUNDLE_API = 'bundle';
 const BUNDLE_API_LIST = `${BUNDLE_API}/list`;
 const BUNDLE_API_COUNT = `${BUNDLE_API}/count`;
-
+const RESOURCE_API = 'resource';
+const RESOURCE_API_LIST = RESOURCE_API;
 /*
 {
   "099a30a6-b707-4df8-b4dd-7149f25658b7": {
@@ -60,7 +68,7 @@ const BUNDLE_API_COUNT = `${BUNDLE_API}/count`;
 
   TO this format:
    {
-      id: 'bundle02', name: 'Another Bundle', revision: 3, task: 'UPLOAD', status: 'UPLOADING', progress: 63, mode: 'PAUSED'
+      id: 'bundle02', name: 'Another Bundle', revision: 3, task: 'UPLOAD', status: 'IN_PROGRESS', progress: 63, mode: 'PAUSED'
    },
  */
 function fetchAll() {
@@ -78,19 +86,27 @@ function convertBundleApiListToBundles(apiBundles) {
     const {
       mode, metadata, dbl, store
     } = apiBundle;
-    let task = dbl.currentRevision === 0 ? 'UPLOAD' : 'DOWNLOAD';
-    let status = dbl.currentRevision === 0 ? 'DRAFT' : 'COMPLETED';
+    let task = dbl.currentRevision === '0' ? 'UPLOAD' : 'DOWNLOAD';
+    let status = dbl.currentRevision === '0' ? 'DRAFT' : 'NOT_STARTED';
     if (mode === 'store') {
       const { history } = store;
       const historyReversed = history.slice().reverse();
-      const action = historyReversed.find((event) => event.type === 'updateStore');
-      task = action.message.toUpperCase();
-      status = 'COMPLETED';
+      const eventUpdateStore = historyReversed.find((event) => event.type === 'updateStore');
+      if (eventUpdateStore && eventUpdateStore.message && eventUpdateStore.message === 'download') {
+        task = eventUpdateStore.message.toUpperCase();
+        const indexOfUpdateStoreDownload = historyReversed.indexOf(eventUpdateStore);
+        const indexOfDownloadResources = historyReversed.findIndex((event) => event.type === 'executeTask' && event.message === 'downloadResources');
+        if (indexOfUpdateStoreDownload < indexOfDownloadResources) {
+          status = 'COMPLETED';
+        }
+      }
     }
     return {
       id: bundleId,
       name: metadata.name,
       revision: dbl.currentRevision,
+      dblId: dbl.id,
+      medium: dbl.medium,
       task,
       status,
     };
@@ -145,5 +161,59 @@ function handleResponse(response) {
   }
 
   return response.json();
+}
+
+function handleTextResponse(response) {
+  if (!response.ok) {
+    return Promise.reject(response.statusText);
+  }
+
+  return response.text();
+}
+
+function getManifestResourcePaths(bundleId) {
+  const requestOptions = {
+    method: 'GET',
+    headers: authHeader()
+  };
+  const url = `${dblDotLocalConfig.getHttpDblDotLocalBaseUrl()}/${BUNDLE_API}/${bundleId}/manifest-resource`;
+  return fetch(url, requestOptions)
+    .then(handleResponse);
+}
+
+function downloadResources(bundleId) {
+  return bundleAddTasks(bundleId, '<downloadResources/>');
+}
+
+function bundleAddTasks(bundleId, innerTasks) {
+  const requestOptions = {
+    method: 'POST',
+    headers: { ...authHeader(), 'Content-Type': 'application/x-www-form-urlencoded' },
+    body: `xml=<tasks> ${innerTasks} </tasks>`
+  };
+  const url = `${dblDotLocalConfig.getHttpDblDotLocalBaseUrl()}/${BUNDLE_API}/${bundleId}/add-tasks`;
+  return fetch(url, requestOptions)
+    .then(handleTextResponse);
+}
+
+
+function getResourcePaths(bundleId) {
+  const requestOptions = {
+    method: 'GET',
+    headers: authHeader()
+  };
+  const url = `${dblDotLocalConfig.getHttpDblDotLocalBaseUrl()}/${BUNDLE_API}/${bundleId}/${RESOURCE_API_LIST}`;
+  return fetch(url, requestOptions)
+    .then(handleResponse);
+}
+
+/*
+ * Downloader.download('https://download.damieng.com/fonts/original/EnvyCodeR-PR7.zip',
+ *  'envy-code-r.zip', (bytes, percent) => console.log(`Downloaded ${bytes} (${percent})`));
+ */
+function requestSaveResourceTo(selectedFolder, bundleId, resourcePath, progressCallback) {
+  const url = `${dblDotLocalConfig.getHttpDblDotLocalBaseUrl()}/${BUNDLE_API}/${bundleId}/${RESOURCE_API}/${resourcePath}`;
+  const targetPath = path.join(selectedFolder, resourcePath);
+  return download(url, targetPath, progressCallback, authHeader());
 }